Create Issue and Workflow upgrade tasks
